Added CH macroses to all libc headers

+++ libc/ctype/isdigit.c
@@ -7,6 +7,6 @@
 */
 #include <ctype.h>
 
-uint8_t isdigit(char c){
+int isdigit(char c){
 	return ((c>='0') && (c<='9'));
 }
