+++ libc/ctype/isprint.c
@@ -1,5 +1,5 @@
 #include <ctype.h>
 
-uint8_t isprint(int c){
+int isprint(int c){
 	return c >= 0x20 && c<= 0x7E;
 }
