+++ libc/ctype/isspace.c
@@ -1,5 +1,5 @@
 #include <ctype.h>
 
-uint8_t isspace(int c){
+int isspace(int c){
 	return c==' ' || c == '\n' || c == '\t' || c == '\v' || c == '\f' || c=='\r';
 }
