+++ libc/ctype/isxdigit.c
@@ -1,5 +1,5 @@
 #include <ctype.h>
 
-uint8_t isxdigit(int c){
+int isxdigit(int c){
 	return (isdigit(c) || (c >= 'A' && c <= 'F') || (c >= 'a' && c <= 'f'));
 }
